% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_name.R
\name{hello_name}
\alias{hello_name}
\title{Hello <Name> - A Javascript to R example (accepting args this time)}
\usage{
hello_name(name)
}
\description{
Hello <Name> - A Javascript to R example (accepting args this time)
}
